fixing meta handlers tests
